changed preprocessing structure
