+++ sklego/preprocessing/__init__.py
@@ -1,0 +1,6 @@
+from .intervalencoder import IntervalEncoder
+from .randomadder import RandomAdder
+from .patsytransformer import PatsyTransformer
+from .pandas import ColumnSelector, PandasTypeSelector, ColumnDropper
+from .projections import InformationFilter, OrthogonalTransformer
+from .repeatingbasis import RepeatingBasisFunction
